# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CoursesList API
  version: "1.0.0"
  summary: API systemu obsługi szkoleń — kursy, kursanci, firmy, zaświadczenia, dzienniki szkoleń.
  description: |
    Specyfikacja dla aplikacji integrujących się z API serwer-serwer za pomocą klucza API.

    **Przeczytaj najpierw `INTEGRATION.md`** z tego samego katalogu — opisuje reguły, których
    nie da się wyrazić w samym schemacie (semantyka PATCH jako pełnego nadpisania, formaty dat
    bez strefy czasowej, niespójności nazw pól, pułapki przy tworzeniu zaświadczeń).

    Konwencje:
    - Każda udana odpowiedź JSON ma kopertę `{"data": ...}`; błędy mają `{"error": {"code", "message"}}`.
    - Każda operacja wymaga zakresu podanego w `x-required-scope` (brak = wystarczy ważny klucz).
    - Operacje z `x-requires-admin: true` wymagają dodatkowo, by konto serwisowe klucza miało rolę
      administratora (`role = 1`) — sam zakres nie wystarczy.
    - Ciała żądań są odrzucane (400), jeśli zawierają pole spoza schematu.

    Wygenerowano z kodu źródłowego (`api/internal/*`) na gałęzi `api_and_webhook`.
servers:
  - url: "{baseUrl}/api/v1"
    variables:
      baseUrl:
        default: http://localhost:8081
        description: >-
          Adres Go API. Lokalnie `http://localhost:8081`; w `docker-compose.prod.yml` API nasłuchuje na
          porcie 8082. Integracje powinny wołać Go API bezpośrednio, nie przez proxy Nuxt.
security:
  - apiKey: []
...
paths:
...
  /companies:
    get:
      tags: [companies]
      operationId: listCompanies
      summary: Lista firm
      description: >-
        Sortowanie: nazwa, miasto. `search` dopasowuje fragment (bez rozróżniania wielkości liter)
        w nazwie, mieście, NIP-ie lub osobie kontaktowej.
      x-required-scope: companies:read
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Firmy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [companies]
      operationId: createCompany
      summary: Utworzenie firmy
      x-required-scope: companies:write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyWrite"
      responses:
        "201":
          description: Utworzona firma.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyDetailsResponse"
        "400":
          description: >-
            `invalid request body` albo niepoprawny NIP: `nip validation error: nip must contain exactly 10 digits`,
            `nip validation error: nip contains invalid characters`, `nip validation error: invalid nip checksum`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: "Firma o tym NIP-ie już istnieje (`company with this NIP already exists`)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: >-
        Klucz API w nagłówku `Authorization: Bearer clk_...`. Klucze wystawia administrator w aplikacji
        webowej (Administracja → Klucze API). Klucz działa w imieniu przypisanego konta serwisowego i
        tylko w zakresie nadanych uprawnień (`x-required-scope`).
  parameters:
...
    Search:
      name: search
      in: query
      description: Fraza wyszukiwania; białe znaki na brzegach są obcinane, pusta = brak filtra.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      description: Maksymalna liczba wyników. Poza zakresem 1–100 → 400 `incorrect limit value`.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  responses:
    BadRequest:
      description: >-
        Błędne żądanie — np. `invalid request body` (brak wymaganego pola, pole spoza schematu,
        zła data), błędny identyfikator w ścieżce, `incorrect limit value`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unauthorized:
      description: >-
        Brak lub nieważny klucz: `missing credentials`, `invalid or expired api key`.
        Klucz mógł zostać unieważniony, wygasnąć albo jego konto usunięto.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: >-
        Klucz jest ważny, ale nie ma uprawnień: `api key is missing the <zakres> scope`
        albo `admin access required` (konto serwisowe nie jest administratorem).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Błąd serwera (`internal_error`). Treść komunikatu nie jest stabilna.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [bad_request, unauthorized, invalid_credentials, forbidden, not_found, conflict, internal_error]
            message:
              type: string
              description: >-
                Komunikat po angielsku. Komunikaty wymienione w opisach operacji są stabilne i można na nich
                opierać logikę; pozostałe traktuj jako informacyjne.
...
    CompanyListItem:
      type: object
      required: [id, name, city, nip, contactPerson, telephone]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        city:
          type: string
        nip:
          type: string
        contactPerson:
          type: [string, "null"]
        telephone:
          type: string
    CompanyListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CompanyListItem"
    CompanyDetails:
      type: object
      required: [id, name, street, city, zipcode, nip, email, contactPerson, telephone, note, expiryNotificationsEnabled, expiryNotificationEmail]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        street:
          type: string
        city:
          type: string
        zipcode:
          type: string
        nip:
          type: string
        email:
          type: [string, "null"]
        contactPerson:
          type: [string, "null"]
        telephone:
          type: string
        note:
          type: [string, "null"]
        expiryNotificationsEnabled:
          type: boolean
          description: Czy firma dostaje e-mailowe przypomnienia o wygasających zaświadczeniach.
        expiryNotificationEmail:
          type: [string, "null"]
          description: Adresy odbiorców przypomnień rozdzielone przecinkami (bez spacji).
    CompanyDetailsResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/CompanyDetails"
    CompanyWrite:
      type: object
      additionalProperties: false
      description: >-
        Wymagane pola nie mogą być puste po obcięciu białych znaków. NIP jest walidowany (10 cyfr, suma
        kontrolna), a myślniki i spacje są usuwane — w odpowiedzi i w bazie NIP to same cyfry.
      required: [name, street, city, zipcode, nip, telephone]
      properties:
        name:
          type: string
          minLength: 1
        street:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        zipcode:
          type: string
          minLength: 1
        nip:
          type: string
          minLength: 1
        email:
          type: [string, "null"]
        contactPerson:
          type: [string, "null"]
        telephone:
          type: string
          minLength: 1
        note:
          type: [string, "null"]
        expiryNotificationsEnabled:
          type: boolean
          default: false
        expiryNotificationEmail:
          type: [string, "null"]
          description: >-
            Lista maks. 10 adresów e-mail rozdzielonych przecinkami (format RFC 5322, np.
            `a@x.pl, Jan <b@x.pl>`). Duplikaty są usuwane. Niepoprawny adres lub więcej niż 10 → 400.
...